Rename ports.py to interfaces.py for better clarity
- Rename core/artifacts/ports.py to core/artifacts/interfaces.py
- Update all imports in core/artifacts/__init__.py and parse_catalog use case
- 'interfaces' better describes the abstract protocol definitions
- All tests passing with updated imports
