+++ build_stream/core/artifacts/__init__.py
@@ -30,7 +30,7 @@ from .exceptions import (
     ArtifactValidationError,
 )
 from .entities import ArtifactRecord
-from .ports import ArtifactStore, ArtifactMetadataRepository
+from .interfaces import ArtifactStore, ArtifactMetadataRepository
 
 __all__ = [
     "ArtifactKey",
